requirements.txt
numpy==1.23.4
pandas==1.5.0
pillow==9.2.0
matplotlib==3.6.1